Add a new parallel type, the SSHPool.
The SSHPool distributes jobs over different nodes using ssh.
Additionally it has framework code to sync files to/from remote
filesystems over ssh.

Note that, by default, no files are synchronized. Hence, targets should
explicitly support this new pool type and will often not work
out-of-the-box. Exceptions to this are nodes with shared filesystems, or
(in some cases) when the same directory structure exists on all nodes.
For the latter to work, the project/infrastructure should be cloned in
identical places on the filesystem of each node, and the build command
should be run manually on each node.