webservers: Refactor directories and scripts.
This commit pulls apart the bash scripts generated by the python code
for (remotely) running webservers, to be more generic. They are now
pieced together by generic getters on webservers (e.g., how to
start/stop the particular server). Additionally, some of the logic is
moved to the python side, such as generating the index.html file.

This commit also refactors the different directories used by all
webserver code. This in part facilitaties the above change, but more
importantly makes it more explicit where directories are valid. This is
important for (future) setups and code that do not have a single shared
filesystem (e.g., when we want to run remotely over SSH later).

The code now has 3 directories: stagedir, rundir and logdir:
 - stagedir is local to the node where you run setup.py from, and is
   populated by all files that the webserver will need (e.g., configs).
   This whole directory is later copied to each node as its rundir.
 - rundir is the directory local to each node (usually in /tmp) where
   all configs live, and originates from stagedir. This directory is
   guaranteed to be local and low-latency (e.g., not going over nfs or
   so). As such, all input and output files should go here.
 - logdir is the local directory where results should go, such as the
   output of wrk and cpu measurements. This directory will later be
   copied to the origin node as entry in the results/ dir.