Static type checking compatibility.
In addition to adding type annotations in many places, this commit also
changes some of the data structures and logic used to be compatible with
static type checking. The most notable of these changes is the removal
of Namespace objects, creating dedicated dataclasses for the context.
This commit also introduces specialized datastructured that are returned
by run and pool.run, instead of inserting data directly into subprocess
types.

This commit may *break* some uses:
 - The insertion of arbitrary fields into the context. If data must be
   passed between instances and targets, this should be documented and
   added to the context, so other targets can also benefit from this.
   Some fields that were used in sporadic places: extra_libs,
   lib_ldflags, cppflags (not cxxflags)
 - Similarly to the above, benchmark-specific fields in the context are
   removed. If any of such uses is still desired, it should be
   considered whether this field can be made more general, or
   a specialized mechanism should be introduced to pass target-specific
   context-data.
   Some fields that were used sporadically: coptimize (and similar),
   spec*_exclude_benchmarks, arch.